+++ NEWS.md
@@ -12,7 +12,8 @@ Version 1.8.0.9000
 
 * [#1007](https://github.com/tidyverse/lubridate/issues/1007) Internal lubridate formats are no longer propagated to stamp formater.
 * `train` argument in `parse_date_time` now takes effect. It was previously ignored.
-* [#1004](https://github.com/tidyverse/lubridate/issues/1004) Fix c.POSIXct and c.Date on empty single POSIXct and Date vectors
+* [#1004](https://github.com/tidyverse/lubridate/issues/1004) Fix `c.POSIXct` and `c.Date` on empty single POSIXct and Date vectors.
+* [#1013](https://github.com/tidyverse/lubridate/issues/1013) Fix c(`POSIXct`,`POSIXlt`) heterogeneous concatenation.
 * [#1002](https://github.com/tidyverse/lubridate/issues/1002) Parsing only with format `j` now works on numeric inputs.
 
 Version 1.8.0